Fix changing layer not cleaning up now blank areas
Since the now blank areas didn't have images, they wouldn't be blanked,
and show the wrong layer. We can't remove the tiles (tried that) because
then when changing back to the layer that did have the images, they
wouldn't get added back. The simple solution is to just hide them.

Also update the version of Leaflet to fix a couple bugs and update the
stylesheets with the better browser support.