Generate Space Exploration star maps
This was quite a journey to get working, so details below. TL;DR: Space
Exploration renders star maps on a Surface, so it is possible to render
them with this program, but dirty hacks are needed to make it work.

The first issue and discovery is that the "starmap-1" surface is NOT the
starmap for the first star. The "1" is for the player with index 1. Each
player gets their own surface where the mod renders whatever starmap
they are currently looking at. Seems crazy, but since each force can
discover different parts of a solar system, the mod needs a way to show
each force something different. Smart.

So the first hurdle was to exclude these "not a surface" Surfaces from
the normal mapping. The solution was to use Zone.from_surface(). If the
mod returns a Zone object, then it is a valid surface players can walk
on. Otherwise, it is a mod internal surface.

The second hurdle is to map these star maps. Since it is just one
Surface, the mod will render different graphics on it depending on what
GUI button was clicked. We can hook into the mod and force it to render
what we want with MapView.start_system_map() and
MapView.start_interstellar_map(). Then we just keep screenshotting the
surface until we have all the maps.

But there is one final (multi-part) hurdle. Like mentioned, the star
maps are just rendered graphics on a surface. They need to finish
rendering before we can screenshot. Thankfully, there is a function for
this: https://lua-api.factorio.com/1.1.61/LuaGameScript.html#LuaGameScript.set_wait_for_screenshots_to_finish
But this did not work. The reason why is fairly complicated.

Factorio renders per tick (which is why the frame rate is capped to the
UPS). Telling the mod to render the map multiple times in one tick is
pointless; the last command is the only thing that we will see rendered.
Along with that, the set_wait_for_screenshots_to_finish() function does
not work like the docs say. Thankfully, I found the thread where
@Rseding91 added it: https://forums.factorio.com/viewtopic.php?p=399852#p399852

> When set it causes the queued screenshots to be force flushed to disk
> at the end of the current tick.

This was the clue we needed. Screenshots don't cause another render,
they take the next render to happen and capture it to disk. And the
function doesn't block, it just forces screenshots to flush at the end
of the tick.

So the solution here is to render and screenshot one star map, then wait
one tick, then do the next etc. Hence the convoluted mess below. Create
a list of all the stars we want to map, then every tick we map one of
them.