Reset coordinates to origin when switching surfaces
This was an obvious problem when switching to a space platform, as they
are much smaller than a planet. If you had panned far from the origin,
you would be completely lost and not know where to find the space
platform.

In Space Exploration, it made some sense not to reset the view, as
planets and their "orbit" surface were 1:1 mapped, and space elevators
would link in the exact same spot on the paired surface. But Space Age
does not do anything of the sort, and so resetting makes much more
sense.