Fix deleted surfaces being removed from map
While the images obviously still existed, a deleted surface would not
show up in the input surfaces data file, which was used as the source of
truth. Now we merge the previous version with the new version, so any
deleted surfaces will still show up.

The issue is that Factorio reuses surface indexes. This can make the map
confusing. There is no good way around this. Either we modify how
Factorio works, which would require a mod and break achievements, or we
ID surfaces by name, which would break anytime someone renamed a space
platform.